Added the survey and all the javascript files
I add the mainly finished survey page with all of the needed JavaScript files as well as a new CSS file for the survey page. I am unsure what the survey page looks like format wise as I couldn't get it to let me use the live preview again so I did it all in text editor. In that, it all looked accurate and worked through. I had to break up the JavaScript files based on functions as it wouldn't work for some reason all in one file.